+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 0.0.1 ]
+## [ v0.1.0 ]
+### Changed
+- Cleaning up python syntax, adding annotations, and updating comments.
+
+### Added
+- Unit Tests for logger.
+- Deprecation warnings methods / classes no longer being used.
+- Ability to show / hide fields: 'timestamps' and 'source ( class / method name )'.
+
+### Removed
+- Deprecating methods for name changes and for refactoring
+
+## [ v0.0.1 ]
 ### Added
 - Changelog to mark the start of versioning.